this is the home page
home_page.php is where the user after successfully log in, it shows the option that the user can do. I already linked the book link to the "book.php" and the log out link to the "logout.php". the view and cancel needs to be linked w the pages.